Conflicts:
	src/airline.c

+++ include/airline.h
@@ -2,10 +2,15 @@
 #define _AIRLINE_
 
 #include "common.h"
+#include "plane.h"
+#include <sys/wait.h>
+#include <sys/types.h>
 #include <string.h>
+#include <unistd.h>
+
 
 typedef struct {
-	long id;
+	int id;
 	int planeCount;
 	int* targetedCities;
 } Airline;
